Fixed incorrect desiredOffset value
Fixed bug where both desiredOffset is an HTMLElement and userOptions is set and is an HTMLElement. Bug was that if partially scrolled through the element, the desiredOffset would produce incorrect values and cause the scrolling to "bounce" and not stay in the location where it ought to have scrolled to.